+++ tests/testthat/test-dictionary.R
@@ -79,3 +79,33 @@ test_that("vec_match works with empty data frame", {
   )
   expect_equal(out, vec_na(integer(), 3))
 })
+
+
+# splits ------------------------------------------------------------------
+
+test_that("can split empty vector", {
+  out <- vec_split(integer(), character())
+
+  expect_s3_class(out, "data.frame")
+  expect_equal(out$key, character())
+  expect_equal(out$val, list_of(.ptype = integer()))
+})
+
+test_that("split data frame with data frame", {
+  df <- data.frame(x = c(1, 1, 2), y = c(1, 1, 1))
+  out <- vec_split(df, df)
+
+  expect_s3_class(out, "data.frame")
+  expect_equal(out$key, data.frame(x = c(1, 2), y = c(1, 1)))
+  expect_equal(out$val, list_of(
+    data.frame(x = c(1, 1), y = c(1, 1)),
+    data.frame(x = 2, y = 1)
+  ))
+})
+
+test_that("x and by must be same size", {
+  expect_error(
+    vec_split(1:3, 1:2),
+    "same size"
+  )
+})